Initial version if traebeler with a rough skeleton of the main worker and processors.
The traefik package handles all the interaction with the traefik v2 REST API and exposes a limited set of simple functions. Within the processing package different processors can be implemented which will be triggered as soon as traefik was successfully queried. Currently only a single processor is supported but it might become handy supporting multiple ones on a update.

The data which is forwarded to a processor is always the full result set. There is no caching or diffing up to now.